Rework projectile state authority
An object with the ProjectManager script will now attempt to request the state authority of an object whenever it spawns.

This method does appear to somewhat work. However, with the clients that aren't the main one, their projectiles will appear to freeze for a short amount of time while transferring state authority. In addition, projectiles spawned to close to objects may not have any effect on those objects.

Also included commented out code for an alternative solution that involves  requesting the state authority of all objects within a certain radius around the projectiles and players.