Add explosive weapon and projectile
The player can now select the explosive weapon and spawn explosive projectiles, which explode upon impact.

However, the projectile currently does not have any initial motion, and it does not utilise the object pooling system yet. In addition, the explosion power may be too weak, or the current objects may be too heavy.